Moved the sort selection, added drop shadow on dropdown Fixes 29
